Initial Task Manager Project Push
This project is very similar to the prvious Finance Tracker project,
with a few twists in how the data is being saved.

Project Purpose: Define a secure system capable of keeping personal task
information. This system will also allow full user control over
available tasks and their completion status.

Implementation:
ShelfService: Python has a module names shelf. The project requirements
ask to store key values of username/password alongside other data. We
can create a method to utilize the shelf (which creates a database
capable of being accessed like a dictionary) to store the username as a
key, with the password, and all relevant tasks behind that. I will use
the same file for storing tasks associated with the user and their
password

UserClass: more of an abstract class that is meant to force the
UIManager (which we will talk about soon) to have a username and a
password attribute. While it really shouldn't have this attribute, I
want to practice with private variables that don't allow getters or
setters, so I'm going to do that

CryptographyService: The project asks to hash the data. I want a service
that does this and only this

Task: Holds the attributes for a task id, it's description, and it's
status. Getters and setters for all attributes

StateManager: This will hold the state for the current user, which
includes the task dictionary (acts as a cache/validation). This will be
optimistically loaded with data once a task is created.

UIManager: Handles a couple of scenarios
1. Registration - If the user wants to register, we check to see if the
   username already exists, and if not then we add it upon successful
   registration with a password
2. Sign In - If you want to sign back into a current session, input
   username and password
3. Add a task - Will need to confirm the description, but besides that
   the task pretty much makes itself. After the task is created, we just
   add it to the dictionary.
4. View Tasks - Will display a table with the appropriate tasks
5. Mark a task complete - the user will select an id from the table of
   tasks, then the task will be marked from Pending to Complete. The
   shelf will be subsequently updated
6. Delete a task - the user will select an id from the table of tasks,
   then the task will be removed from the table. The dictionary in the
   shelf will be subsequently updated
7. Logout - a message for goodbye will be issued, and the program will
   save one last time for good measure

Once again, I am going to challenge myself to only use the keyboard for
this project, and build from the floor up, only testing the project once
I have what I believe to be the final product. Wish me luck