Shelf Service and new ThreadService
I've created a new file for a class that will be the ThreadService... I
decided I didn't want to clutter up the Shelf at all, and my
ThreadFactory methods will be inside of a ThreadService rather than
attempting to put it in my shelf. Strategically, this makes far more
sense, since I had planned to have a factory in the first place.

I'm a bit uncertain now though where I'm going to put the functions that
need to be threaded.

Let's say that someone inputs their username succesfully when trying to
sign in. I can go ahead and fetch that user data while they are typing
in their password and process it. Now the question is, do I put this
combination of shelf->process data->back to user input?

process data is eseentially just mapping the tasks to a dictionary
format using their id's and then giving it back. So what I'm curious
about is, where do I put the async queue to fetch that data, and at what
stage in my code do I insert such logic to push to the async queue?

I want the factory method, but I also don't want to force it. It's and
interesting dilemma, so I might have to go back to my design here in a
second and try and redistribute some things to go into their proper
place