ThreadService Initialization
I was making the process a little harder in my head than it actually
should be. Turns out, shelf can actually save your class types, so when
I pull something off the shelf, I won't have to do any processing to get
them to be a Task object. How neat is that!

So, the issue of where certain building blocks go is now eliminated. I
can create my threadFactory and not have to worry so much about where to
put my async queue... it just goes in the Factory!

There's definitely some logic flaws to that in general... chances are I
shouldn't actually be putting a queue in the factory itself. But, I
think it will be ok for this project, and I will look more into system
design to see how it's done at a later date. I can't go down every
alleyway that comes into sight, I have to pick and choose the ones most
relevant to me at the given time. That one is not so relevant to me
right now