Creating the StateManager
I am now creating the state manager for the whole project. All that does
is hold the values of the current User, whoever that may be. It is
important to do this, since we have a logout feature, and upon further
relection, the "logout" feature should switch to another user/register for another user if you so desire.