Everything Almost Worked First Try
So, there were a couple of errors made, but not something that wasn't
easily fixable. The main one was that I forgot to put brackets around
my data constructor function, so the function reference itself was being
saved to the shelf and not the actual data. Whoopsie!

But other than that, it seemed like things went pretty smoothly! Not a
lot of changes in this commit besides spacing some things out with \n's
and fixing that one error... I would call this a huge success!

I will have to make a branching strategy for some thing... I'll want to
comment things out a bit and get the threading to work. But for now, I
think that this will do quite nicely as a first iteration